+++ metta/map/load_random_from_index.py
@@ -3,7 +3,7 @@ import random
 from metta.map.load import Load
 from mettagrid.util import file as file_utils
 
-from .scene import SceneCfg
+from .types import SceneCfg
 
 
 class LoadRandomFromIndex(Load):
