+++ metta/map/scenes/auto.py
@@ -1,7 +1,6 @@
 import numpy as np
 
 from metta.map.node import Node
-from metta.map.scene import AreaWhere, ChildrenAction
 from metta.map.scenes.bsp import BSPLayout
 from metta.map.scenes.make_connected import MakeConnected
 from metta.map.scenes.mirror import Mirror
@@ -9,7 +8,7 @@ from metta.map.scenes.random import Random
 from metta.map.scenes.random_objects import RandomObjects
 from metta.map.scenes.random_scene import RandomScene, RandomSceneCandidate
 from metta.map.scenes.room_grid import RoomGrid
-from metta.map.types import MapGrid
+from metta.map.types import AreaWhere, ChildrenAction, MapGrid
 from metta.map.utils.random import FloatDistribution, IntDistribution, sample_int_distribution
 from metta.util.config import Config
